Revert "書類登録画面のリンクをdocuments/editからdocuments/newに変更"
This reverts commit 97a14576ac588420747607a8a0b2278ac0769f4f.
