+++ app/(main)/documents/edit/[[...id]]/page.tsx
@@ -1,9 +1,9 @@
 'use client';
 
 import { DocumentFormEditor } from '@/components/3_organisms/documents/document-form-editor';
-import { useDocumentSave } from '@/hooks/useDocumentSave';
-import { useRouter, useSearchParams } from 'next/navigation';
-import React, { useEffect } from 'react';
+import type { DocumentFormData } from '@/validations/document-validation';
+import { useSearchParams } from 'next/navigation';
+import React from 'react';
 
 interface PageProps {
   params: Promise<{
@@ -13,7 +13,6 @@ interface PageProps {
 
 export default function DocumentEditPage({ params }: PageProps) {
   const searchParams = useSearchParams();
-  const router = useRouter();
 
   // URLパラメータからドキュメントIDを取得
   const resolvedParams = React.use(params);
@@ -22,41 +21,52 @@ export default function DocumentEditPage({ params }: PageProps) {
   // URLクエリパラメータからフォルダIDを取得
   const folderId = searchParams.get('folder') || searchParams.get('folderId');
 
-  // 新規作成の場合は新規作成ページにリダイレクト
-  useEffect(() => {
-    if (!documentId) {
-      const redirectUrl = folderId ? `/documents/new?folder=${folderId}` : '/documents/new';
-      router.replace(redirectUrl);
-    }
-  }, [documentId, folderId, router]);
-
-  // 新規作成の場合は何も表示しない（リダイレクト中）
-  if (!documentId) {
-    return null;
-  }
-
-  // 保存処理フックを使用
-  const { saveDocument } = useDocumentSave({ documentId, folderId });
-
   // 編集モードの場合のモックデータ読み込み
   // 実際のアプリケーションではAPIからデータを取得
-  const initialDocument = {
-    id: documentId,
-    title: 'サンプル書類タイトル',
-    content: '<p>ここに書類の内容が入ります。</p>',
-    category: '議事録',
-    description: 'サンプルの説明文です。',
-    status: 'draft' as const,
-    tags: '会議,報告',
-    createdAt: new Date(),
-    updatedAt: new Date(),
+  const initialDocument = documentId
+    ? {
+        id: documentId,
+        title: 'サンプル書類タイトル',
+        content: '<p>ここに書類の内容が入ります。</p>',
+        category: '議事録',
+        description: 'サンプルの説明文です。',
+        status: 'draft' as const,
+        tags: '会議,報告',
+        createdAt: new Date(),
+        updatedAt: new Date(),
+      }
+    : undefined;
+
+  // 保存処理
+  const handleSave = async (data: {
+    formData: DocumentFormData;
+    content: string;
+    attachedFile?: File;
+  }) => {
+    try {
+      // 実際のアプリケーションではAPIに送信
+      const saveData = {
+        documentId,
+        folderId, // フォルダIDも保存データに含める
+        ...data,
+      };
+      // console.log('Saving document:', saveData);
+
+      // 保存処理をシミュレート
+      await new Promise((resolve) => setTimeout(resolve, 500));
+
+      return true;
+    } catch (error) {
+      console.error('Save failed:', error);
+      return false;
+    }
   };
 
   return (
     <DocumentFormEditor
       documentId={documentId}
       initialDocument={initialDocument}
-      onSave={saveDocument}
+      onSave={handleSave}
     />
   );
 }
