+++ components/2_molecules/documents/document-toolbar.tsx
@@ -10,7 +10,6 @@ import {
   SelectValue,
 } from '@/components/ui/select';
 import {
-  FileText,
   Filter,
   FolderPlus,
   Grid3X3,
@@ -19,6 +18,7 @@ import {
   Shield,
   SortAsc,
   Upload,
+  FileText,
 } from 'lucide-react';
 import Link from 'next/link';
 import type React from 'react';
@@ -55,7 +55,7 @@ export const DocumentToolbar: React.FC<DocumentToolbarProps> = ({
       {/* Top row - Actions */}
       <div className="flex items-center justify-between">
         <div className="flex items-center gap-2">
-          <Link href="/documents/new">
+          <Link href="/documents/edit">
             <Button className="bg-carebase-blue hover:bg-carebase-blue-dark">
               <FileText className="h-4 w-4 mr-2" />
               新規書類
